a big mess
this starts working towards integrating bevy, bevy_ratatui,
bevy_ratatui_camera, and ratatui_image. I mostly have the ratatui stuff
working, the camera -> image pipeline is not yet functional but it isn't
broken either.

I think the camera render is probably due to my lack of bevy knowledge,
but in principal it should be possible to pipe the image into
ratatui_image and use the protocol implementation there. If that
continues to fail, I'll go ahead and start work on creating a strategy
on the bevy_ratatui_camera crate, as I think there may be some weirdness
in porting it's `camera_image` component all the way to ratatui's end.

The picker setup seems to work, logging logs to widget, and overall I
think this is a good checkpoint for what I'm aiming towards.